Fixed bug
Before: when measuring current it would take ten measurements in the span of a second
This was done for each ina because the loop was inside the read_ina method.
Obviously it added up to 4 seconds no matter what. Which is longer than 3.

So I moved the loop into measure() method. Also, ina objects are created only once
when creating a class object now. Basically got rid of bad code.